move TextController to base-tool
