+++ ui/base-tool/src/main/java/org/cru/godtools/base/tool/ui/controller/TextController.kt
@@ -1,20 +1,19 @@
-package org.cru.godtools.tract.ui.controller
+package org.cru.godtools.base.tool.ui.controller
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
-import org.cru.godtools.base.tool.ui.controller.BaseController
-import org.cru.godtools.tract.databinding.TractContentTextBinding
+import org.cru.godtools.base.tool.databinding.ToolContentTextBinding
 import org.cru.godtools.xml.model.Text
 
 internal class TextController private constructor(
-    private val binding: TractContentTextBinding,
+    private val binding: ToolContentTextBinding,
     parentController: BaseController<*>
 ) : BaseController<Text>(Text::class, binding.root, parentController) {
     @AssistedInject internal constructor(@Assisted parent: ViewGroup, @Assisted parentController: BaseController<*>) :
-        this(TractContentTextBinding.inflate(LayoutInflater.from(parent.context), parent, false), parentController)
+        this(ToolContentTextBinding.inflate(LayoutInflater.from(parent.context), parent, false), parentController)
 
     @AssistedFactory
     interface Factory : BaseController.Factory<TextController>
