+++ ui/base-tool/src/main/res/layout/tool_content_text.xml
@@ -7,8 +7,8 @@
     <TextView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginLeft="@dimen/tract_content_margin_horizontal"
-        android:layout_marginRight="@dimen/tract_content_margin_horizontal"
+        android:layout_marginLeft="@dimen/tool_content_margin_horizontal"
+        android:layout_marginRight="@dimen/tool_content_margin_horizontal"
         android:saveEnabled="false"
         android:text="@{model}"
         android:textIsSelectable="true" />
