+++ ui/base-tool/src/main/res/values/styles_content.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="tool_content_margin_horizontal">16dp</dimen>
+</resources>
